// generated/addon/angular/src/app/api-models/cieb-stc-mapping.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebStcMapping {
  seqStcId: number;
  svcTypeCode: string;
  stcCovgType: string;
  detailStc: string;
  ruleGroupId: number;
  ruleLevel: number;
  medDefCode: string;
  medDefDet: string;
  medDefOrder: string;
  shortRow: string;
  ebQtyQual: string;
  ebQty: string;
  benPosCode: string;
  msgText: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  ppacaMsgInd: string;
}